Support templatized `CFBundleShortVersionString`
from Info.plist instead of hard-coded. There was this app that did this, and it looks like Sparkle (or a fork of it) understands what to do. Without this patch, the script explodes and doesn't process any versions. It probably should rescue and inform the user of version parse failure too.